+++ src/main/java/com/raoulvdberge/refinedstorage/container/ContainerCrafter.java
@@ -1,21 +1,21 @@
 package com.raoulvdberge.refinedstorage.container;
 
+import com.raoulvdberge.refinedstorage.container.slot.SlotBase;
 import com.raoulvdberge.refinedstorage.tile.TileCrafter;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.Slot;
 import net.minecraft.item.ItemStack;
-import net.minecraftforge.items.SlotItemHandler;
 
 public class ContainerCrafter extends ContainerBase {
     public ContainerCrafter(TileCrafter crafter, EntityPlayer player) {
         super(crafter, player);
 
         for (int i = 0; i < 9; ++i) {
-            addSlotToContainer(new SlotItemHandler(crafter.getPatternItems(), i, 8 + (18 * i), 20));
+            addSlotToContainer(new SlotBase(crafter.getPatternItems(), i, 8 + (18 * i), 20));
         }
 
         for (int i = 0; i < 4; ++i) {
-            addSlotToContainer(new SlotItemHandler(crafter.getUpgrades(), i, 187, 6 + (i * 18)));
+            addSlotToContainer(new SlotBase(crafter.getUpgrades(), i, 187, 6 + (i * 18)));
         }
 
         addPlayerInventory(8, 55);