+++ src/main/java/com/raoulvdberge/refinedstorage/container/slot/SlotOutput.java
@@ -2,11 +2,10 @@ package com.raoulvdberge.refinedstorage.container.slot;
 
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.items.IItemHandler;
-import net.minecraftforge.items.SlotItemHandler;
 
 import javax.annotation.Nonnull;
 
-public class SlotOutput extends SlotItemHandler {
+public class SlotOutput extends SlotBase {
     public SlotOutput(IItemHandler inventory, int id, int x, int y) {
         super(inventory, id, x, y);
     }